remove extra container component
