+++ client/app/page.tsx
@@ -1,6 +1,6 @@
 // Home page
 
-import ControlPanel from '@/components/control-panel';
+import KeyboardSelected from '@/components/keyboard-selected';
 import KeyboardFull from '@/components/keyboard-full';
 import Header from '@/components/layout/header';
 
@@ -11,7 +11,7 @@ export default function Home(): JSX.Element {
 
 			<KeyboardFull />
 
-			<ControlPanel />
+			<KeyboardSelected />
 		</div>
 	);
 }
