+++ client/components/keyboard-selected.test.tsx
@@ -1,40 +1,60 @@
-import { render, screen } from '@testing-library/react';
+import { render, screen, within } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
 
 import { KeyboardOptionsProvider } from '@/context/keyboard-options-context';
-import type { FullNote } from '@/types/keyboard-option-types';
 import KeyboardSelected from './keyboard-selected';
-
-const mockFullNoteOptions: FullNote[] = [
-	'C4',
-	'D♭4',
-	'D4',
-	'E♭4',
-	'E4',
-	'F4',
-	'G♭4',
-	'G4',
-	'A♭4',
-	'A4',
-	'B♭4',
-	'B4',
-];
 
 describe('Selected Keyboard', () => {
 	beforeEach(() => {
 		render(
 			<KeyboardOptionsProvider>
-				<KeyboardSelected fullNotes={mockFullNoteOptions} />
+				<KeyboardSelected />
 			</KeyboardOptionsProvider>
 		);
 	});
 
-	it('renders the selected keyboard div', () => {
-		const selectedKeyboard = screen.getByRole('group', {
-			name: 'Selected keyboard',
+	it('renders the selected keyboard', () => {
+		const keyboardSelected = screen.getByRole('region', {
+			name: 'Selected Keyboard: audio controls and keys',
 		});
 
-		expect(selectedKeyboard).toBeInTheDocument();
+		expect(keyboardSelected).toBeInTheDocument();
+	});
+
+	it('renders the audio controls div', () => {
+		const audioControls = screen.getByRole('group', {
+			name: 'Audio controls',
+		});
+
+		expect(audioControls).toBeInTheDocument();
+	});
+
+	it('rearranges the 13 notes correctly', () => {
+		const octave = screen.getByRole('group', {
+			name: 'Octave for C4',
+		});
+
+		const noteButtons = within(octave).getAllByRole('button');
+		const noteTexts = noteButtons.map((button) =>
+			// remove whitespace
+			button.textContent?.trim().replace(/\s+/g, '')
+		);
+
+		expect(noteTexts).toEqual([
+			'C4',
+			'D♭4',
+			'D4',
+			'E♭4',
+			'E4',
+			'F4',
+			'G♭4',
+			'G4',
+			'A♭4',
+			'A4',
+			'B♭4',
+			'B4',
+			'C5',
+		]);
 	});
 
 	it('renders the default octave with the correct selection of keys', () => {
